+++ src/parser_frontend.mli
@@ -24,8 +24,6 @@ val parse_ocaml : string -> signature
     Raise Not_found if file does not exist. Raise Syntax_error if there is an
     OCaml syntax error. *)
 
-val parse_ocaml_lb : Lexing.lexbuf -> signature
-
 val parse_gospel : filename:string -> signature -> string -> Uast.s_signature
 (** [parse_gospel sig_list module_name] parses the GOSPEL attributes and
     integrates them in the corresponding OCaml signatures. *)